New Sensors, New Tiles
colors are represented in RGB format with leading #
eg.: '#RGB' from 0 to F, '#000' to '#FFF'

- ColoredTile.py
- ColoredTileSensor.py
- HoleSensor.py

+++ environment/Tile.py
@@ -6,5 +6,5 @@ class Tile(object):
     outline = ''
     width = 0
 
-    def __init__(self, vertexes: List[int]):
-        self.vertexes: List[int] = vertexes
+    def __init__(self, vertexes: List[float]):
+        self.vertexes: List[float] = vertexes
